# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vadsa",
    "description": "Swedish speech recognition with KlangAI/pianissimo-sv, served like vLLM.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "description": "Like vLLM: an empty 200 once the model is loaded. No auth.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "The model is loaded."
          },
          "503": {
            "description": "Loading."
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "summary": "List Models",
        "operationId": "list_models_v1_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelList"
                }
              }
            }
          },
          "401": {
            "description": "The API key is missing or wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/audio/transcriptions": {
      "post": {
        "summary": "Create Transcription",
        "operationId": "create_transcription_v1_audio_transcriptions_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Audio in any format ffmpeg reads."
                  },
                  "model": {
                    "type": "string",
                    "description": "The served model name."
                  },
                  "language": {
                    "type": "string",
                    "enum": [
                      "sv",
                      "auto"
                    ],
                    "description": "Only Swedish is served; auto means sv.",
                    "default": "sv"
                  },
                  "response_format": {
                    "type": "string",
                    "enum": [
                      "json",
                      "text",
                      "verbose_json"
                    ],
                    "default": "json"
                  },
                  "timestamp_granularities[]": {
                    "items": {
                      "type": "string",
                      "enum": [
                        "word",
                        "segment"
                      ]
                    },
                    "type": "array",
                    "description": "What verbose_json includes; other formats ignore it.",
                    "default": [
                      "segment"
                    ]
                  },
                  "prompt": {
                    "type": "string",
                    "description": "Accepted and ignored."
                  },
                  "temperature": {
                    "type": "number",
                    "description": "Accepted and ignored."
                  }
                },
                "type": "object",
                "required": [
                  "file",
                  "model"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Transcription"
                    },
                    {
                      "$ref": "#/components/schemas/TranscriptionVerbose"
                    }
                  ],
                  "title": "Response Create Transcription V1 Audio Transcriptions Post"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "The request or the audio is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "The API key is missing or wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`model` is not the served model.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "The upload is larger than allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The model is loading or the server is busy; retry after Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorInfo": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "param": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "message",
          "type",
          "param",
          "code"
        ],
        "title": "ErrorInfo"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "ModelCard": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "object": {
            "type": "string",
            "const": "model",
            "title": "Object",
            "default": "model"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "owned_by": {
            "type": "string",
            "const": "vadsa",
            "title": "Owned By",
            "default": "vadsa"
          },
          "model_type": {
            "type": "string",
            "const": "transcription",
            "title": "Model Type",
            "default": "transcription"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created"
        ],
        "title": "ModelCard"
      },
      "ModelList": {
        "properties": {
          "object": {
            "type": "string",
            "const": "list",
            "title": "Object",
            "default": "list"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ModelCard"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ModelList"
      },
      "Transcription": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "Transcription"
      },
      "TranscriptionSegment": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "start": {
            "type": "number",
            "title": "Start"
          },
          "end": {
            "type": "number",
            "title": "End"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "start",
          "end",
          "text"
        ],
        "title": "TranscriptionSegment"
      },
      "TranscriptionVerbose": {
        "properties": {
          "task": {
            "type": "string",
            "const": "transcribe",
            "title": "Task",
            "default": "transcribe"
          },
          "language": {
            "type": "string",
            "const": "sv",
            "title": "Language",
            "default": "sv"
          },
          "duration": {
            "type": "number",
            "title": "Duration"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "segments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TranscriptionSegment"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segments"
          },
          "words": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TranscriptionWord"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Words"
          }
        },
        "type": "object",
        "required": [
          "duration",
          "text"
        ],
        "title": "TranscriptionVerbose"
      },
      "TranscriptionWord": {
        "properties": {
          "word": {
            "type": "string",
            "title": "Word"
          },
          "start": {
            "type": "number",
            "title": "Start"
          },
          "end": {
            "type": "number",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "word",
          "start",
          "end"
        ],
        "title": "TranscriptionWord"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}